+++ hail/src/main/scala/is/hail/expr/types/physical/PCanonicalStruct.scala
@@ -4,12 +4,16 @@ import is.hail.annotations.{CodeOrdering, Region, UnsafeInserter}
 import is.hail.asm4s.Code
 import is.hail.expr.ir.{EmitMethodBuilder, SortOrder}
 import is.hail.expr.types.BaseStruct
-import is.hail.expr.types.virtual.Type
 import is.hail.utils.{ArrayBuilder, fatal, plural, prettyIdentifier}
 import org.apache.spark.sql.Row
 import is.hail.utils._
 
 object PCanonicalStruct {
+  private val requiredEmpty = PCanonicalStruct(Array.empty[PField], true)
+  private val optionalEmpty = PCanonicalStruct(Array.empty[PField], false)
+
+  def empty(required: Boolean = false): PStruct = if (required) requiredEmpty else optionalEmpty
+
   def apply(required: Boolean, args: (String, PType)*): PStruct =
     PCanonicalStruct(args
       .iterator
@@ -17,12 +21,13 @@ object PCanonicalStruct {
       .map { case ((n, t), i) => PField(n, t, i) }
       .toFastIndexedSeq,
       required)
+
+  def apply(args: (String, PType)*): PStruct =
+    PCanonicalStruct(false, args:_*)
 }
 
 final case class PCanonicalStruct(fields: IndexedSeq[PField], required: Boolean = false) extends PStruct with PCanonicalBaseStruct {
   assert(fields.zipWithIndex.forall { case (f, i) => f.index == i })
-
-  def copy(fields: IndexedSeq[PField] = this.fields, required: Boolean = this.required): PStruct = PCanonicalStruct(fields, required)
 
   val types: Array[PType] = fields.map(_.typ).toArray
 
@@ -31,6 +36,8 @@ final case class PCanonicalStruct(fields: IndexedSeq[PField], required: Boolean 
     fatal(s"cannot create struct with duplicate ${plural(duplicates.size, "field")}: " +
       s"${fieldNames.map(prettyIdentifier).mkString(", ")}", fieldNames.duplicates())
   }
+
+  def copy(fields: IndexedSeq[PField] = this.fields, required: Boolean = this.required): PStruct = PCanonicalStruct(fields, required)
 
   override def truncate(newSize: Int): PStruct =
     PCanonicalStruct(fields.take(newSize), required)
@@ -94,7 +101,7 @@ final case class PCanonicalStruct(fields: IndexedSeq[PField], required: Boolean 
           })
 
         case None =>
-          val (insertedFieldType, fieldInserter) = PCanonicalStruct(Array.empty[PField], required).unsafeInsert(typeToInsert, path.tail)
+          val (insertedFieldType, fieldInserter) = PStruct.empty().unsafeInsert(typeToInsert, path.tail)
 
           (appendKey(key, insertedFieldType), { (region, offset, rvb, inserter) =>
             rvb.startStruct()
@@ -127,7 +134,7 @@ final case class PCanonicalStruct(fields: IndexedSeq[PField], required: Boolean 
     assert(fieldIdx.contains(key))
     val index = fieldIdx(key)
     if (fields.length == 1)
-      PCanonicalStruct(Array.empty[PField], required)
+      PCanonicalStruct.empty()
     else {
       val newFields = Array.fill[PField](fields.length - 1)(null)
       for (i <- 0 until index)
@@ -149,14 +156,12 @@ final case class PCanonicalStruct(fields: IndexedSeq[PField], required: Boolean 
 
 
   def rename(m: Map[String, String]): PStruct = {
-    val newFieldsBuilder = new ArrayBuilder[PField]()
-    var idx = 0
+    val newFieldsBuilder = new ArrayBuilder[(String, PType)]()
     fields.foreach { fd =>
       val n = fd.name
-      newFieldsBuilder += PField(m.getOrElse(n, n), fd.typ, idx)
-      idx += 1
+      newFieldsBuilder += (m.getOrElse(n, n) -> fd.typ)
     }
-    PCanonicalStruct(newFieldsBuilder.result(), required)
+    PCanonicalStruct(newFieldsBuilder.result(): _*)
   }
 
   def ++(that: PStruct): PStruct = {
@@ -165,19 +170,7 @@ final case class PCanonicalStruct(fields: IndexedSeq[PField], required: Boolean 
     if (overlapping.nonEmpty)
       fatal(s"overlapping fields in struct concatenation: ${ overlapping.mkString(", ") }")
 
-    PCanonicalStruct(required, fields.map(f => (f.name, f.typ)) ++ that.fields.map(f => (f.name, f.typ)):_*)
-  }
-
-  def identBase: String = "tuple"
-
-  override def pyString(sb: StringBuilder): Unit = {
-    sb.append("struct{")
-    fields.foreachBetween({ field =>
-      sb.append(prettyIdentifier(field.name))
-      sb.append(": ")
-      field.typ.pyString(sb)
-    }) { sb.append(", ")}
-    sb.append('}')
+    PCanonicalStruct(fields.map(f => (f.name, f.typ)) ++ that.fields.map(f => (f.name, f.typ)): _*)
   }
 
   override def _pretty(sb: StringBuilder, indent: Int, compact: Boolean) {
@@ -205,7 +198,7 @@ final case class PCanonicalStruct(fields: IndexedSeq[PField], required: Boolean 
   }
 
   def select(keep: IndexedSeq[String]): (PStruct, (Row) => Row) = {
-    val t = PCanonicalStruct(required, keep.map { n =>
+    val t = PCanonicalStruct(keep.map { n =>
       n -> field(n).typ
     }: _*)
 
@@ -220,7 +213,7 @@ final case class PCanonicalStruct(fields: IndexedSeq[PField], required: Boolean 
     selectFields(fieldNames.filter(!names.contains(_)))
 
   def typeAfterSelect(keep: IndexedSeq[Int]): PStruct =
-    PCanonicalStruct(required, keep.map(i => fieldNames(i) -> types(i)):_*)
+    PCanonicalStruct(keep.map(i => fieldNames(i) -> types(i)): _*)
 
   override val fundamentalType: PStruct = {
     val fundamentalFieldTypes = fields.map(f => f.typ.fundamentalType)
